Remove modal class, in favor of all modal component
